requirements.txt
google-cloud-secret-manager
google-cloud-functions
openai